+++ src/dataManager/OrderData.cpp
@@ -1,7 +1,8 @@
 #include "OrderData.h"
 
 OrderData::OrderData(OrderTypes::OrderType type,
-                     Exchange::ExchangeName exchange, long price, double volume)
+                     Exchange::ExchangeName exchange, long long price, double
+                                                                      volume)
     : orderType(type),
       exchange(exchange),
       orderPrice(price),
@@ -16,4 +17,4 @@ const Exchange::ExchangeName OrderData::getExchange() const { return exchange; }
 
 const double OrderData::getOrderVolume() const { return orderVolume; }
 
-const long OrderData::getOrderPrice() const { return orderPrice; }
+const long long OrderData::getOrderPrice() const { return orderPrice; }
